Added ProjectFiles.zip for the html download.
Please install the CSS and javascript. Please don't rename them. (if you want to, you may have to change the HTML) And also if you don't install or download them then the HTML won't be styled and if the javascript is not included then the Codes In the YourProfile.html won't work properly because the Tetris game code may include pure javascript.